+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- `XRPClient` requires a new parameter in it's constructor that identifies the network it is attached to.
+
 ## 3.0.4 - March 24, 2020
 
 This release contains minor fixes for exception handling in PayID.